seperate column reducer
